+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ﻿# Changelog
 
-## v1.9.0 (unreleased)
+## v1.9.0
 
 - Introduce default version setting to DurableTaskClient and expose to orchestrator ([#393](https://github.com/microsoft/durabletask-dotnet/pull/393))
 - Add support for local credential types in DTS libraries ([#396](https://github.com/microsoft/durabletask-dotnet/pull/396))